requirements.txt
Django==4.2
django-ninja==0.21.0
environs==9.5.0
selectolax==0.3.12
